split tournament models into seperate files
